Crate forest_cid

Re-exports

pub use multihash;

Modules

json

Structs

Cid
Content identifier for any Ipld data. This Cid consists of a version, a codec (or serialization) protocol and a multihash (hash of the Ipld data). Cids allow for hash linking, where the Cids are used to resolve any arbitrary data over a network or from local storage.

Prefix
Prefix represents all metadata of a CID, without the actual content.

Enums

Code
Multihash generation codes for the Filecoin protocol. This is not an exhausting list of codes used, just the ones used to generate multihashes.

Error
Error types

Version
The version of the CID.

Constants

DAG_CBOR
Cbor Cid codec.

FIL_COMMITMENT_SEALED
Sealed commitment Cid codec.

FIL_COMMITMENT_UNSEALED
Unsealed commitment Cid codec.

POSEIDON_BLS12_381_A1_FC1
Multihash code for Poseidon BLS replica commitments.

RAW
Raw Cid codec. This represents data that is not encoded using any protocol.

SHA2_256_TRUNC254_PADDED
Multihash code for Sha2 256 trunc254 padded used in data commitments.

Functions

new_from_cbor
Constructs a cid with bytes using default version and codec

new_from_prefix
Create a new CID from a prefix and some data.

Type Definitions

Multihash
A Multihash with the same allocated size as the Multihashes produces by this derive.